update new parametric programming examples
