+++ examples/parametric_programming/Part_1_LearnToOptimize_tutorial.py
@@ -1,17 +1,21 @@
 """
-Solve the parametric Rosenbrock problem, formulated as the NLP using Neuromancer toolbox:
-minimize     (1-x)^2 + a*(y-x^2)^2
-subject to   (p/2)^2 <= x^2 + y^2 <= p^2
-             x>=y
+Learning to optimize parametric nonlinear programming problem (pNLP) using Neuromancer.
 
-problem parameters:             a, p
-problem decition variables:     x, y
+Formulation of the parametric Rosenbrock problem:
+
+    minimize     (1-x)^2 + a*(y-x^2)^2
+    subject to   (p/2)^2 <= x^2 + y^2 <= p^2
+                 x>=y
+
+    problem parameters:             a, p
+    problem decition variables:     x, y
 
 https://en.wikipedia.org/wiki/Rosenbrock_function
 """
 
 import torch
 import torch.nn as nn
+import numpy as np
 import neuromancer.slim as slim
 import matplotlib.pyplot as plt
 import matplotlib.patheffects as patheffects
@@ -32,6 +36,8 @@ if __name__ == "__main__":
     """
     # # #  Dataset
     """
+    # We synthetically sample a distribution of the optimization problem parameters
+    # Each element in the set of sampled parameters represents one instance of the problem
     data_seed = 408
     np.random.seed(data_seed)
     nsim = 5000  # number of datapoints: increase sample density for more robust results
@@ -75,26 +81,41 @@ if __name__ == "__main__":
     plt.show(block=True)
 
     """
-    # # #  mpNLP primal solution map architecture
+    # # #  pNLP primal solution map architecture
     """
-    # define neural architecture for the solution map
+    # wrap pytorch callable concatenating problem parameters/features (a, p)
+    # into symbolic representation via the Node class:
+    # features(a, p) -> xi
+    xi = lambda a, p: torch.cat([a, p], dim=-1)
+    features = Node(xi, ['a', 'p'], ['xi'], name='features')
+    # define neural architecture for the trainable solution map
     func = blocks.MLP(insize=2, outsize=2,
                     bias=True,
                     linear_map=slim.maps['linear'],
                     nonlin=nn.ReLU,
                     hsizes=[80] * 4)
-    # define symbolic solution map with concatenated features (problem parameters)
-    xi = lambda a, p: torch.cat([a, p], dim=-1)
-    features = Node(xi, ['a', 'p'], ['xi'], name='features')
+    # wrap neural net into symbolic representation of the solution map via the Node class:
+    # sol_map(xi) -> x
     sol_map = Node(func, ['xi'], ['x'], name='map')
 
     """
-    # # #  mpNLP objective and constraints formulation in Neuromancer
+    # # #  pNLP variables, objective, and constraints formulation in Neuromancer
+    
+    variable is basic symbolic abstraction in Neuromancer
+       x = variable("variable_name")                      (instantiates new variable)  
+    variable construction supports:
+       algebraic expressions:     x**2 + x**3 + 5     (instantiates new variable)  
+       slicing:                   x[:, i]             (instantiates new variable)  
+       pytorch callables:         torch.sin(x)        (instantiates new variable)  
+       constraints definition:    x <= 1.0            (instantiates Constraint object) 
+       objective definition:      x.minimize()        (instantiates Objective object) 
+    to visualize computational graph of the variable use x.show() method          
     """
-    # variables
+
+    # define decision variables
     x = variable("x")[:, [0]]
     y = variable("x")[:, [1]]
-    # sampled parameters
+    # problem parameters sampled in the dataset
     p = variable('p')
     a = variable('a')
 
@@ -125,7 +146,7 @@ if __name__ == "__main__":
     # plt.show(block=True)
 
     """
-    # # #  mpNLP problem solution in Neuromancer
+    # # #  pNLP problem solution in Neuromancer
     """
     optimizer = torch.optim.AdamW(problem.parameters(), lr=0.001)
     # define trainer
@@ -144,7 +165,7 @@ if __name__ == "__main__":
         eval_metric="dev_loss",
     )
 
-    # Train mpLP solution map
+    # Train problem solution map
     best_model = trainer.train()
     best_outputs = trainer.test(best_model)
     # load best model dict
@@ -212,7 +233,7 @@ if __name__ == "__main__":
     plt.setp(cg3.collections,
              path_effects=[patheffects.withTickedStroke()], alpha=0.7)
 
-    # Solution to mpNLP via Neuromancer
+    # Solution to pNLP via Neuromancer
     datapoint = {'a': torch.tensor([[a]]), 'p': torch.tensor([[p]]),
                  'name': 'test'}
     model_out = problem(datapoint)
